% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseIndex.R
\name{cumgmean}
\alias{cumgmean}
\title{Cumulative geometric mean}
\usage{
cumgmean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of positive values}

\item{na.rm}{should missing values be removed (default to TRUE)}
}
\value{
the cumulative geometric mean of x
}
\description{
Cumulative geometric mean
}
\examples{
gmean(c(0.5, 0.5))
}
