% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseIndex.R
\name{gmean}
\alias{gmean}
\title{Geometric mean}
\usage{
gmean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of positive values}

\item{na.rm}{should missing values be removed (default to TRUE)}
}
\value{
the geometric mean of x
}
\description{
Geometric mean
}
\examples{
gmean(c(0.5, 0.5))
}
