% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm}
\alias{norm}
\alias{normL1}
\alias{normL2}
\alias{normLG}
\title{L1-norm, L2-norm or LG-norm of a vector of numerics}
\usage{
normL1(vec)

normL2(vec)

normLG(vec, grp)
}
\arguments{
\item{vec}{vector of numeric values}

\item{grp}{vector describing the groups}
}
\value{
the L1-norm, L2-norm or LG-norm of vec
}
\description{
L1-norm, L2-norm or LG-norm of a vector of numerics
}
\examples{
x <- c(-0.1, 1, 0.5)
g <- c(1, 1, 2)
normL1(x) # = 1.6
normL2(x) # ~= 1.12
normLG(x, g) # ~= 1.5
}
