% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{print.geigen}
\alias{print.geigen}
\title{Print a \code{geigen} object}
\usage{
\method{print}{geigen}(x, ...)
}
\arguments{
\item{x}{an object of class \code{geigen}, typically produced by \code{geigen} (from the \code{GSVD} package)}

\item{...}{additional arguments, currently unused}
}
\value{
\code{x}, invisibly; called for its side effect of printing a summary to the console
}
\description{
Print a \code{geigen} object
}
