% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{print.gplssvd}
\alias{print.gplssvd}
\title{Print a \code{gplssvd} object}
\usage{
\method{print}{gplssvd}(x, ...)
}
\arguments{
\item{x}{an object of class \code{gplssvd}, typically produced by \code{\link{gplssvd}}}

\item{...}{additional arguments, currently unused}
}
\value{
\code{x}, invisibly; called for its side effect of printing a summary to the console
}
\description{
Print a \code{gplssvd} object
}
