% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{print.gsvd}
\alias{print.gsvd}
\title{Print a \code{gsvd} object}
\usage{
\method{print}{gsvd}(x, ...)
}
\arguments{
\item{x}{an object of class \code{gsvd}, typically produced by \code{gsvd} (from the \code{GSVD} package) or \code{\link{sparseGSVD}}}

\item{...}{additional arguments, currently unused}
}
\value{
\code{x}, invisibly; called for its side effect of printing a summary to the console
}
\description{
Print a \code{gsvd} object
}
