% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj}
\alias{proj}
\alias{projL1}
\alias{projL2}
\alias{projL1L2}
\alias{projLG}
\alias{projLGL2}
\alias{projOrth}
\alias{projPos}
\alias{projL1L2_then_projOrth}
\alias{projLGL2_then_projOrth}
\alias{projOrth_then_projL1L2}
\alias{projOrth_then_projLGL2}
\alias{projOrth_then_projPos_then_projL1L2}
\alias{projOrth_then_projPos_then_projLGL2}
\title{L1-, L2-, L1L2- or LG- projection operators applied to a vector of numerics}
\usage{
projL1(vec, rds)

projL2(vec, rds = 1)

projL1L2(vec, rds, method = "regular")

projLG(vec, rds, grp)

projLGL2(vec, rds, grp, method = "regular")

projOrth(vec, OrthSpace)

projPos(vec)

projL1L2_then_projOrth(vec, rds, grp = NULL, OrthSpace, itermax, eps)

projLGL2_then_projOrth(vec, rds, grp, OrthSpace, itermax, eps)

projOrth_then_projL1L2(vec, rds, grp = NULL, OrthSpace, itermax, eps)

projOrth_then_projLGL2(vec, rds, grp, OrthSpace, itermax, eps)

projOrth_then_projPos_then_projL1L2(
  vec,
  rds,
  grp = NULL,
  OrthSpace,
  itermax,
  eps
)

projOrth_then_projPos_then_projLGL2(vec, rds, grp, OrthSpace, itermax, eps)
}
\arguments{
\item{vec}{vector of numeric values}

\item{rds}{radius}

\item{method}{which algorithm to use, either "regular" (default) or "fast"}

\item{grp}{vector describing the groups}

\item{OrthSpace}{matrix defining the orthogonal space}

\item{itermax}{the maximum number of POCS iterations}

\item{eps}{precision for the POCS iterations}
}
\value{
the L1-, L2-, L1L2- or LG- projection of a vector of numerics
}
\description{
L1-, L2-, L1L2- or LG- projection operators applied to a vector of numerics
}
\examples{
x <- c(-0.1, 1, 0.5)
M <- rbind(c(-1, 1, 0) / sqrt(2))
r <- 1
g <- c(1, 1, 2)
projL1(vec = x, rds = r)$x # = (0, 0.75, 0.25)
projL2(vec = x)$x # ~= (-0.1, 0.9, 0.4)
projL1L2(vec = x, rds = r)$x # = (0, 1, 0)
projLG(vec = x, rds = r, grp = g)$x # ~= (-0.1, 0.7, 0.2)
projLGL2(vec = x, rds = r, grp = g)$x # ~= (-0.1, 1.0, 0)
projOrth(vec = x, OrthSpace = M)$x # = (-0.55, 0.55, 0.50)
}
