% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{prox}
\alias{prox}
\alias{proxL1}
\alias{proxL2}
\alias{proxLG}
\alias{proxLG2}
\title{L1-, L2- or LG- proximal mapping of a vector of numerics}
\usage{
proxL1(vec, lambda)

proxL2(vec, lambda)

proxLG(vec, lambda, grp)

proxLG2(vec, lambda, grp)
}
\arguments{
\item{vec}{vector of numeric values}

\item{lambda}{proximal mapping parameter}

\item{grp}{vector describing the groups}
}
\value{
the L1-, L2- or LG- proximal mapping of a vector of numerics
}
\description{
L1-, L2- or LG- proximal mapping of a vector of numerics
}
\examples{
x <- c(-0.1, 1, 0.5)
lamb <- 0.5
g <- c(1, 1, 2)
proxL1(x, lamb) # = (0, 0.5, 0)
proxL2(x, lamb) # ~= (-0.1, 0.6, 0.3)
proxLG(x, lamb, c(1, 2, 3)) # = (0, 0.5, 0)
proxLG(x, lamb, c(1, 1, 1)) # ~= (-0.1, 0.6, 0.3)
proxLG(x, lamb, g) # ~= (-0.1, 0.5, 0)
}
