% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseGEIGEN.R
\name{sparse_geigen}
\alias{sparse_geigen}
\title{Sparse generalized eigenvalue decomposition}
\usage{
sparse_geigen(
  X,
  W,
  k = 0,
  R = 2L,
  init = NULL,
  seed = NULL,
  rds = rep(1, R),
  grp = NULL,
  orthogonality = "loadings",
  OrthSpace = NULL,
  projPriority = "orth",
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{X}{a square, semi-positive symmetric data matrix to decompose}

\item{W}{\bold{W}eights -- the constraints applied to the matrix and thus the eigen vectors.}

\item{k}{total number of components to return though the full variance will still be returned (see \code{d_full}). If 0, the full set of components are returned.}

\item{c}{-- the sparsity constraint.}

\item{tol}{default is \code{sqrt(.Machine$double.eps)}. A tolerance level for eliminating effectively zero (small variance), negative, imaginary eigen/singular value components.}

\item{symmetric}{if \code{X} is symmetric, set as TRUE. See \code{\link{eigen}}.}
}
\value{
A list with eight elements:
\item{d_full}{A vector containing the singular values of X above the tolerance threshold (based on eigenvalues).}
\item{l_full}{A vector containing the eigen values of X above the tolerance threshold (\code{tol}).}
\item{d}{A vector of length \code{min(length(d_full), k)} containing the retained singular values of X}
\item{l}{A vector of length \code{min(length(l_full), k)} containing the retained eigen values of X}
\item{v}{Eigenvectors. Dimensions are \code{ncol(X)} by k.}
\item{q}{Generalized eigenvectors. Dimensions are \code{ncol(X)} by k.}
\item{fj}{Component scores. Dimensions are \code{ncol(X)} by k.}
}
\description{
\code{sparse_sparse_geigen} takes in a positive definite matrix of constraints (\code{W}), and a positive real number setting the required amount of sparsity in the resulting pseudo-eigen vectors.
}
\examples{

## (Metric) Multidimensional Scaling
data(wine, package="GSVD")
D <- as.matrix(dist(wine$objective))
masses <- rep(1/nrow(D), nrow(D))
Xi <- diag(nrow(D)) - ( rep(1,nrow(D)) \%o\% masses )
S <- Xi \%*\% (-(D^2) / 2) \%*\% t(Xi)
mds.res_sparse_geigen <- sparse_geigen(S)

## Principal components analysis: "covariance"
cov_X <- as.matrix(cov(wine$objective))
cov_pca.res_sparse_geigen <- sparse_geigen(cov_X)

## Principal components analysis: "correlation"
cor_X <- as.matrix(cor(wine$objective))
cor_pca.res_sparse_geigen <- sparse_geigen(cor_X)

}
\seealso{
\code{\link{tolerance_eigen}}, \code{\link{gsvd}} and \code{\link{gplssvd}}
}
\author{
Derek Beaton
}
\keyword{multivariate}
