% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseEIGENpos.R
\name{sparseEIGENpos}
\alias{sparseEIGENpos}
\title{Constrained eigen-value decomposition of a symmetric matrix with an additionnal positive constraint}
\usage{
sparseEIGENpos(
  X,
  k = 2L,
  init = "svd",
  seed = NULL,
  rds = rep(1, k),
  grp = NULL,
  orthogonality = "loadings",
  OrthSpace = NULL,
  projPriority = "sparsity",
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{X}{a symmetric square (data) matrix;}

\item{k}{the desired rank of the singular decomposition;}

\item{init}{How to initialize the algorithm, Default: 'svd'}

\item{seed}{a random seed for result reproducibility; if NULL (the default), no random seed will be used}

\item{rds}{The radiuses (radii?) (>0) of the L1 or LG constraint; one for each dimension;}

\item{grp}{vector describing the groups; default to one group per row}

\item{orthogonality}{whether the orthogonality constraint is applied on the "loadings" (default)}

\item{OrthSpace}{matrix defining the orthogonal space, Default: NULL}

\item{projPriority}{the order in which the projections are applied, Default: 'sparsity'}

\item{itermaxALS}{the maximum number of ALS iterations, Default: 1000}

\item{itermaxPOCS}{the maximum number of POCS iterations, Default: 1000}

\item{epsALS}{precision for ALS, Default: 1e-10}

\item{epsPOCS}{precision for POCS, Default: 1e-10}
}
\value{
Pseudo-eigen vectors and values
}
\description{
Constrained eigen-value decomposition of a symmetric matrix with an additionnal positive constraint
}
\examples{
U <- matrix(rnorm(20), 5, 4)
sparseEIGENpos(U \%*\% t(U))
}
\author{
Vincent Guillemot
}
