% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseGSVD.R
\name{sparseGSVD}
\alias{sparseGSVD}
\title{Sparse Generalized Singular Value Decomposition}
\usage{
sparseGSVD(
  X,
  Y = NULL,
  LW,
  RW,
  LM,
  RM,
  k = 2L,
  tol = .Machine$double.eps,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, k),
  rdsRight = rep(1, k),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  correction4SI = "gsvd",
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10,
  tol.si = .Machine$double.eps
)
}
\arguments{
\item{X}{a (data) matrix;}

\item{Y}{a second (data) matrix; this is optional and is only used for a two-table method, Default: NULL}

\item{LW}{left constraints (weights) applied to X (and thus to the left pseudo-singular vectors)}

\item{RW}{right constraints (weights) applied to X (and thus to the right pseudo-singular vectors)}

\item{LM}{left masses, used to compute the singular vectors' contributions}

\item{RM}{right masses, used to compute the singular vectors' contributions}

\item{k}{the desired rank of the singular decomposition, Default: 0}

\item{tol}{tolerance below which singular values are considered zero, Default: .Machine$double.eps}

\item{init}{How to initialize the algorithm, Default: 'svd'}

\item{initLeft}{how the left pseudo-singular vectors are initialized, Default: NULL}

\item{initRight}{how the right pseudo-singular vectors are initialized, Default: NULL}

\item{seed}{a random seed for result reproducibility; if NULL (the default), no random seed will be used, Default: NULL}

\item{rdsLeft}{The radius (>0) of the
$L_1$ ball for each left vector, Default: rep(1, k)}

\item{rdsRight}{The radius (>0) of the $L_1$ balls for each right vector, Default: rep(1, k)}

\item{grpLeft}{vector describing the groups for the left vectors, Default: NULL}

\item{grpRight}{vector describing the groups for the right vectors, Default: NULL}

\item{orthogonality}{whether the orthogonality constraint is applied on the "loadings" (default), the "scores", or "both"}

\item{OrthSpaceLeft}{matrix defining the orthogonal space for the left vectors, Default: NULL}

\item{OrthSpaceRight}{matrix defining the orthogonal space for the right vectors, Default: NULL}

\item{projPriority}{the order in which the projections are applied, Default: 'orth'}

\item{projPriorityLeft}{the order in which the projections are applied for the left vectors, Default: projPriority}

\item{projPriorityRight}{the order in which the projections are applied for the right vectors, Default: projPriority}

\item{correction4SI}{Correction for the explained variance for sparsity indices, e.g., "gsvd" (no correction), "mca" (Benzecri correction of MCA), or "mfa"}

\item{itermaxALS}{The maximum number of ALS iterations, Default: 1000}

\item{itermaxPOCS}{The maximum number of the POCs iterations, Default: 1000}

\item{epsALS}{Precision in ALS, Default: 1e-10}

\item{epsPOCS}{Precision in POCs, Default: 1e-10}

\item{tol.si}{Tolerance for the computation of the Sparse Index, set by default to .Machine$double.eps}
}
\value{
Pseudo-singular vectors and values
}
\description{
Constrained SVD of a matrix (wrapper of c++ functions).
}
\details{
DETAILS
}
\examples{
X <- matrix(rnorm(20), 5, 4)
sparseSVD(X)
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Vincent Guillemot, Ju-Chi Yu
}
