% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseIndex.R
\name{sparseIndex}
\alias{sparseIndex}
\alias{sparseIndexEigen}
\title{Compute sparse indices on the results of a sparse GSVD}
\usage{
sparseIndex(res.sgsvd, singularValues, correction = "gsvd", tol = 1e-10)

sparseIndexEigen(res.sgevd, eigenValues, correction = "gevd", tol = 1e-10)
}
\arguments{
\item{res.sgsvd}{The result of a sparse Generalized Singular Value Decomposition of a data-matrix, usually obtained with sGSVD::sparseGSVD or one of the companion functions in the SPAFAC package ;}

\item{singularValues}{The singular values of the original data matrix (used to compute the fit of the sparse analysis)}

\item{correction}{The type of correction for proportion of explained variance (r1), e.g., "gsvd" (no correction), "mca", "mfa"}

\item{tol}{a tolerance parameter indicating when a small value should be considered equal to 0}

\item{res.sgevd}{The result of a sparse Generalized EigenValue Decomposition of a data-matrix, usually obtained with sGSVD::sparseGEIGEN}

\item{eigenValues}{The eigenvalues of the original data matrix (used to compute the fit of the sparse analysis)}
}
\value{
various sparsity indices and their components.
}
\description{
Compute sparse indices on the results of a sparse GSVD
}
\examples{
set.seed(2045)
X <- matrix(rnorm(20), 4, 5)
res.svd <- svd(X)
res.sgsvd <- sparseGSVD(X, k = 2L)
sparseIndex(res.sgsvd, res.svd$d)
}
