% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.column.R
\name{unique_column}
\alias{unique_column}
\title{Exclude duplicated columns with a tolerance}
\usage{
unique_column(x, n.round)
}
\arguments{
\item{x}{a marix with columns to screen}

\item{n.round}{to allow tolerance of the difference between values, specify the rounding digit}
}
\value{
The unique columns of \eqn{x}.
}
\description{
Exclude duplicated columns with a tolerance
}
\examples{
x <- matrix(c(1,2,3, 1.001, 2.002, 3.003), nrow = 3, ncol = 2, byrow = FALSE)
unique_column(x, n.round = 2)
}
