% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedSparsityIndex.R
\name{weightedSparsityIndex}
\alias{weightedSparsityIndex}
\title{Calculate Weighted Sparsity Indices from Sparse GSVD Results}
\usage{
weightedSparsityIndex(
  res.sgsvd,
  singularValues,
  correction = "gsvd",
  weight = 0.5,
  tol = 1e-10
)
}
\arguments{
\item{res.sgsvd}{A list containing the results of a sparse Generalized
Singular Value Decomposition (sGSVD) applied to a data matrix.
Typically generated using `sGSVD::sparseGSVD` or similar functions in
the sGSVD package.}

\item{singularValues}{A numeric vector of the singular values from the
original data matrix, used to assess the fit of the sparse GSVD.
These are usually obtained from a standard SVD or GSVD.}

\item{correction}{A string indicating the type of correction to apply when
calculating the proportion of explained variance (`r1`). Options include
`"gsvd"` (no correction), `"mca"`, or `"mfa"`. Default is `"gsvd"`.}

\item{weight}{A numeric value (between 0 and 1) specifying the weight
applied to the fit ratio when combining it with the sparsity measure.
The complement `(1 - weight)` is applied to the sparsity ratio. Default
is `0.5`, meaning equal weight for fit and sparsity.}

\item{tol}{A numeric tolerance value to define when elements should be
considered zero. Any value below this threshold is treated as zero.
Default is `1e-10`.}
}
\value{
A list with three components:
\describe{
  \item{SI}{A numeric vector containing the weighted sparsity indices
    combining fit and sparsity.}
  \item{SIleft}{A numeric vector containing the weighted sparsity indices
    for the left singular vectors (rows).}
  \item{SIright}{A numeric vector containing the weighted sparsity indices
    for the right singular vectors (columns).}
}
}
\description{
This function computes weighted sparsity indices based on the results of a
sparse Generalized Singular Value Decomposition (sGSVD). It combines measures
of sparsity and fit to provide a comprehensive index.
}
\examples{
set.seed(2045)
X <- matrix(rnorm(20), 4, 5)
res.svd <- svd(X)
res.sgsvd <- sparseGSVD(X, k = 2L)
weightedSparsityIndex(
    res.sgsvd = res.sgsvd,
    singularValues = res.svd$d,
    correction = "gsvd",
    weight = 0.4) # (slightly) favors sparsity over fit
}
