% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/als.R
\name{als}
\alias{als}
\title{Alternative least squares method helper for the cGSVD}
\usage{
als(
  X,
  initLeft,
  initRight,
  projLeft,
  projRight,
  rdsLeft,
  rdsRight,
  grpLeft = NULL,
  grpRight = NULL,
  OrthSpaceLeft,
  OrthSpaceRight,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{X}{data matrix on which the ALS is performed}
}
\value{

}
\description{
Alternative least squares method helper for the cGSVD
}
\examples{

}
