% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseIndex.R
\name{compute.fit}
\alias{compute.fit}
\title{Compute r1 (the explained variance) for sparsity indices}
\usage{
compute.fit(d, pseudo.d, J, correction = "gsvd")
}
\arguments{
\item{d}{a vector of singular values}

\item{pseudo.d}{a vector of the pseudo singular values (from sparseGSVD)}

\item{J}{the number of variables (for MCA correction)}

\item{correction}{the type of correction, e.g., "gsvd" (no correction), "mca", "mfa"}
}
\value{
the corrected eigenvalues and tau
}
\description{
Compute r1 (the explained variance) for sparsity indices
}
\examples{

}
