% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseIndex.R
\name{cummean}
\alias{cummean}
\title{Cumulative arithmetic mean}
\usage{
cummean(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of numeric values}
}
\value{
the cumulative arithmetic mean
}
\description{
Cumulative arithmetic mean
}
