% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseEIGENpos.R
\name{sparseEIGENpos}
\alias{sparseEIGENpos}
\title{Constrained eigen-value decomposition of a symmetric matrix with an additionnal positive constraint}
\usage{
sparseEIGENpos(
  X,
  k = 2L,
  init = NULL,
  seed = NULL,
  rds = rep(1, k),
  grp = NULL,
  orthogonality = "loadings",
  OrthSpace = NULL,
  projPriority = "sparsity",
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{X}{a symmetric square (data) matrix;}

\item{k}{the desired rank of the singular decomposition;}

\item{init}{How to initialize the algorithm}

\item{rds}{The radiuses (radii?) (>0) of the L1 or LG constraint; one for each dimension;}

\item{epsPOCS}{$L_1$ ball for each left vector}
}
\value{
Pseudo-eigen vectors and values
}
\description{
Constrained eigen-value decomposition of a symmetric matrix with an additionnal positive constraint
}
\examples{
U <- matrix(rnorm(20), 5, 4)
sparseEIGEN(U \%*\% t(U))
}
\author{
Vincent Guillemot
}
