% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseGSVD.R
\name{sparseGSVD}
\alias{sparseGSVD}
\title{Sparse Generalized Singular Value Decomposition}
\usage{
sparseGSVD(
  X,
  Y = NULL,
  LW,
  RW,
  LM,
  RM,
  k = 0,
  tol = .Machine$double.eps,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, k),
  rdsRight = rep(1, k),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10,
  tol.si = .Machine$double.eps
)
}
\arguments{
\item{X}{a (data) matrix;}

\item{Y}{a second (data) matrix; this is optional and is only used for a two-table method, Default: NULL}

\item{LW}{PARAM_DESCRIPTION}

\item{RW}{PARAM_DESCRIPTION}

\item{k}{the desired rank of the singular decomposition, Default: 0}

\item{tol}{PARAM_DESCRIPTION, Default: .Machine$double.eps}

\item{init}{How to initialize the algorithm, Default: 'svd'}

\item{initLeft}{PARAM_DESCRIPTION, Default: NULL}

\item{initRight}{PARAM_DESCRIPTION, Default: NULL}

\item{seed}{PARAM_DESCRIPTION, Default: NULL}

\item{rdsLeft}{The radius (>0) of the
$L_1$ ball for each left vector, Default: rep(1, k)}

\item{rdsRight}{The radius (>0) of the $L_1$ balls for each right vector, Default: rep(1, k)}

\item{grpLeft}{PARAM_DESCRIPTION, Default: NULL}

\item{grpRight}{PARAM_DESCRIPTION, Default: NULL}

\item{orthogonality}{PARAM_DESCRIPTION, Default: 'loadings'}

\item{OrthSpaceLeft}{PARAM_DESCRIPTION, Default: NULL}

\item{OrthSpaceRight}{PARAM_DESCRIPTION, Default: NULL}

\item{projPriority}{PARAM_DESCRIPTION, Default: 'orth'}

\item{projPriorityLeft}{PARAM_DESCRIPTION, Default: projPriority}

\item{projPriorityRight}{PARAM_DESCRIPTION, Default: projPriority}

\item{itermaxALS}{The maximum number of ALS iterations, Default: 1000}

\item{itermaxPOCS}{The maximum number of the POCs iterations, Default: 1000}

\item{epsALS}{Precision in ALS, Default: 1e-10}

\item{epsPOCS}{Precision in POCs, Default: 1e-10}

\item{tol.si}{Tolerance for the computation of the Sparse Index, set by default to .Machine$double.eps}
}
\value{
Pseudo-singular vectors and values
}
\description{
Constrained SVD of a matrix (wrapper of c++ functions).
}
\details{
DETAILS
}
\examples{
X <- matrix(rnorm(20), 5, 4)
sparseSVD(X)
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Vincent Guillemot, Ju-Chi Yu
}
