% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseIndex.R
\name{sparseIndex}
\alias{sparseIndex}
\title{Compute a sparse index on the results of a sparse GSVD}
\usage{
sparseIndex(X, res.gsvd, tol = 1e-16)
}
\arguments{
\item{X}{the data matrix}

\item{res.gsvd}{the result of a sparse GSVD of X}

\item{tol}{a tolerance parameter indicating when a small value should be considered equal to 0}
}
\value{

}
\description{
Compute a sparse index on the results of a sparse GSVD
}
