% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseSVD.R
\name{sparseSVD}
\alias{sparseSVD}
\title{Constrained SVD of a matrix (wrapper of c++ functions).}
\usage{
sparseSVD(
  X,
  Y = NULL,
  k = 2L,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, k),
  rdsRight = rep(1, k),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)
}
\arguments{
\item{X}{a (data) matrix;}

\item{Y}{a second (data) matrix; this is optional and is only used for a two-table method}

\item{k}{the desired rank of the singular decomposition;}

\item{init}{How to initialize the algorithm}

\item{au}{The radiuses (radii?) (>0) of the
$L_1$ ball for each left vector}

\item{av}{The radiuses (radii)?
(>0) of the $L_1$ balls for each right vector}

\item{itermax}{The maximum number of iterations}

\item{eps}{Precision}
}
\value{
Pseudo-singular vectors and values
}
\description{
Constrained SVD of a matrix (wrapper of c++ functions).
}
\examples{
X <- matrix(rnorm(20), 5, 4)
sparseSVD(X)
}
\author{
Vincent Guillemot
}
