% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseSVD.R
\name{sparseSVD}
\alias{sparseSVD}
\title{Constrained SVD of a matrix.}
\usage{
sparseSVD(
  X,
  Y = NULL,
  k = 2L,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, k),
  rdsRight = rep(1, k),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  compute_sparsity_index = TRUE,
  correction4SI = "gsvd",
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10,
  tol.si = .Machine$double.eps
)
}
\arguments{
\item{X}{a (data) matrix ;}

\item{Y}{a second (data) matrix ; this is optional and is only used for two-table methods ;}

\item{k}{the desired rank of the singular decomposition ;}

\item{init}{how the pseudo-singular vectors are initialized ; default to "svd", meaning that they are initialized with regular singular vectors ; any other value will initialize them randomly ;}

\item{initLeft}{how the left pseudo-singular vectors are initialized ; default to the same value as "init" ;}

\item{initRight}{how the right pseudo-singular vectors are initialized ; default to the same value as "init" ;}

\item{seed}{a random seed for result reproducibility ; if NULL (the default), no random seed will be used ;}

\item{rdsLeft}{a vector of radiuses
(>0) of the $L_1$ or $L_G$ balls for each of the k left vectors ;}

\item{rdsRight}{a vector of radiuses
(>0) of the $L_1$ or $L_G$ balls for each of the k right vectors ;}

\item{grpLeft}{vector describing the groups for the left vectors ; default to one group per row ;}

\item{grpRight}{vector describing the groups for the right vectors ; default to one group per column ;}

\item{orthogonality}{wh the orthogonality constraint will}

\item{itermaxALS}{The maximum number of ALS iterations}

\item{itermaxPOCS}{The maximum number of POCS iterations}

\item{epsALS}{Precision for ALS}

\item{epsPOCS}{Precision for POCS}
}
\value{
Pseudo-singular vectors and values
}
\description{
Constrained SVD of a matrix.
}
\examples{
X <- matrix(rnorm(20), 5, 4)
sparseSVD(X)
}
\author{
Vincent Guillemot
}
